seperate out brewery list item component
