+++ pages/index.tsx
@@ -17,6 +17,7 @@ import DetailsModal from "../components/DetailsModal";
 import SearchIcon from "@mui/icons-material/Search";
 import { LoadingButton } from "@mui/lab";
 import { AcUnit } from "@mui/icons-material";
+import BreweryListItem from "../components/BreweryListItem";
 
 export interface BreweryData {
   id: string;
@@ -34,19 +35,7 @@ export interface BreweryData {
 export default function Home() {
   const [loading, setLoading] = useState(false);
   const [input, setInput] = useState("");
-  const [selectedBrewery, setSelectedBrewery] = useState<BreweryData | null>(
-    null
-  );
   const [breweries, setBreweries] = useState([]);
-  const handleOpen = (brewery: BreweryData) => () => {
-    setSelectedBrewery((selectedCampaign) =>
-      selectedBrewery ? null : brewery
-    );
-  };
-
-  const handleClose = () => {
-    setSelectedBrewery(null);
-  };
 
   const callAPI = async () => {
     setLoading(true);
@@ -61,7 +50,7 @@ export default function Home() {
       const metadata = await metadataResponse.json();
       setBreweries(byCityData);
       console.log(byCityData);
-      console.log(metadata)
+      console.log(metadata);
     } catch (err) {
       console.log(err);
     } finally {
@@ -133,60 +122,10 @@ export default function Home() {
         >
           <List>
             {breweries.map((brewery: BreweryData) => (
-              <>
-                <ListItem key={brewery.id}>
-                  <ListItemText
-                    primary={brewery.name}
-                    secondary={
-                      <>
-                        <Typography>
-                          <Typography
-                            sx={{ display: "inline" }}
-                            fontWeight="500"
-                          >
-                            Address:
-                          </Typography>{" "}
-                          {brewery.street} {brewery.city}, {brewery.state},{" "}
-                          {brewery.postal_code}
-                        </Typography>
-                        <Typography>
-                          <Typography
-                            sx={{ display: "inline" }}
-                            fontWeight="500"
-                          >
-                            Type:{" "}
-                          </Typography>
-                          {brewery.brewery_type}
-                        </Typography>
-                        <Typography>
-                          <Typography
-                            sx={{ display: "inline" }}
-                            fontWeight="500"
-                          >
-                            Website:{" "}
-                          </Typography>
-                          <Link
-                            href={brewery.website_url}
-                            target="_blank"
-                            rel="noopener noreferrer"
-                          >
-                            {brewery.website_url}
-                          </Link>
-                        </Typography>
-                      </>
-                    }
-                  ></ListItemText>
-                  <Button variant="outlined" onClick={handleOpen(brewery)}>
-                    View Details
-                  </Button>
-                  <DetailsModal
-                    open={selectedBrewery?.id === brewery.id}
-                    brewery={brewery}
-                    onClose={handleClose}
-                  />
-                </ListItem>
-                <Divider />
-              </>
+              <BreweryListItem
+                key={brewery.id}
+                brewery={brewery}
+              />
             ))}
           </List>
         </Paper>
